FTK Item ID Filter Builder
This script can take a long list of FTK item ID's and generate an XML filter that can be imported into FTK. Simply place the list of item ID's in a file named "item_ids.txt" in the same directory as the script before running it, and a file titled "denerated_filter.xml" Will be outputted.

I believe this filter will only properly work with/filtering on Item ID's. But I'm sure someone more talented than I can can easily tweak the code to make it work for other values/filter types, or ask ChatGPT to tweak it lol.